requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langsmith==0.0.87
langchain-openai==0.0.5
scikit-learn==1.3.0
xgboost==2.0.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
plotly==5.17.0
scipy==1.11.0
fastapi==0.104.1
uvicorn==0.24.0
click==8.1.7
rich==13.7.0
pytest==7.4.3
pytest-asyncio==0.21.1
tenacity==8.2.3
pydantic==2.5.0
shap==0.43.0
dtaidistance==2.3.11
joblib==1.3.2